feat: fix Values MAX value serialization with additional bit, move classes into separate files
